+++ src/utils/render.js
@@ -1,0 +1,40 @@
+const RenderPosition = {
+  AFTERBEGIN: `afterbegin`,
+  BEFOREEND: `beforeend`
+};
+
+/**
+ * Рендерит шаблон верстки в контейнер
+ * @param  {element} container DOM-элемент рендеринга
+ * @param  {object} component  элемент вставки
+ * @param  {string} place     местоположение в контейнере
+ */
+const render = (container, component, place) => {
+  switch (place) {
+    case RenderPosition.AFTERBEGIN:
+      container.prepend(component.getElement());
+      break;
+    case RenderPosition.BEFOREEND:
+      container.append(component.getElement());
+      break;
+  }
+};
+
+const createElement = (template) => {
+  const newElement = document.createElement(`div`);
+  newElement.innerHTML = template;
+
+  return newElement.firstChild;
+};
+
+const remove = (component) => {
+  component.getElement().remove();
+  component.removeElement();
+};
+
+export {
+  RenderPosition,
+  render,
+  createElement,
+  remove
+};
